module for head to head is now working though I'm now completely sure why
